don't treat integers as hash keys when populating paths
Currently when setting values, if the current reference in
the path is a hash an integer path part is treated as key.

One could argue from symmetry that integers should always be treated
as array references when new components are created, so that using
them as hash keys should result in an error (e.g. returning undef).